+++ cmd/root.go
@@ -2,6 +2,7 @@ package cmd
 
 import (
 	"fmt"
+	"net/http"
 	"os"
 
 	"github.com/dynatrace-oss/dtwiz/pkg/logger"
@@ -13,6 +14,7 @@ import (
 var Version = "dev"
 
 var debugFlag bool
+var verbosityFlag int
 var environmentFlag string
 var accessTokenFlag string
 var platformTokenFlag string
@@ -31,8 +33,11 @@ Set your Dynatrace credentials via environment variables:
 Then use dtwiz commands to analyze and instrument your system.`,
 	PersistentPreRun: func(cmd *cobra.Command, args []string) {
 		debug := debugFlag || os.Getenv("DT_DEBUG") == "true"
-		logger.Init(debug)
-		logger.Debug("Debug mode enabled")
+		logger.Init(debug, verbosityFlag)
+		if logger.Verbosity() > 0 {
+			http.DefaultTransport = logger.NewLoggingTransport(http.DefaultTransport)
+		}
+		logger.Debug("Debug mode enabled", "verbosity", logger.Verbosity())
 	},
 }
 
@@ -67,7 +72,8 @@ func init() {
 		cmd.Help()
 	}
 
-	rootCmd.PersistentFlags().BoolVar(&debugFlag, "debug", false, "enable debug logging (also read from DT_DEBUG=true)")
+	rootCmd.PersistentFlags().BoolVar(&debugFlag, "debug", false, "enable full HTTP request/response debug logging (equivalent to -vv; also read from DT_DEBUG=true)")
+	rootCmd.PersistentFlags().CountVarP(&verbosityFlag, "verbose", "v", "verbose output (-v for request/response summary, -vv for full headers and bodies)")
 	rootCmd.PersistentFlags().StringVar(&environmentFlag, "environment", "", "Dynatrace environment URL (also read from DT_ENVIRONMENT)")
 	rootCmd.PersistentFlags().StringVar(&accessTokenFlag, "access-token", "", "Dynatrace API access token (also read from DT_ACCESS_TOKEN)")
 	rootCmd.PersistentFlags().StringVar(&platformTokenFlag, "platform-token", "", "Dynatrace platform token (dt0s16.*) for AWS installer (also read from DT_PLATFORM_TOKEN)")